requirements.txt
openai
jinja2
weasyprint
python-dotenv
pydantic
